WIP data assem

+++ src/compiler/codegen.h
@@ -74,9 +74,9 @@ struct CG_frag_
     enum {CG_stringFrag, CG_procFrag, CG_dataFrag} kind;
     union
     {
-        struct {Temp_label label; string str; int msize;} stringg;
+        struct {Temp_label label; string str; size_t msize;} stringg;
         struct {S_pos pos; Temp_label label; bool expt; AS_instrList body; CG_frame frame;} proc;
-        struct {Temp_label label; bool expt; int size; CG_dataFragNode init; CG_dataFragNode initLast;} data;
+        struct {Temp_label label; bool expt; size_t size; CG_dataFragNode init; CG_dataFragNode initLast;} data;
     } u;
 };
 struct CG_dataFragNode_
